添加前端代码实现单体部署 解决从分布式模块合并到单体模块的模块冲突问题

+++ colorful-fog/src/main/java/com/ruoyi/colorfulfog/modules/datasource/service/JdbcService.java
@@ -16,7 +16,7 @@ public interface JdbcService {
      *
      * @param id
      */
-    void removeJdbcConnectionPool(Integer id);
+    void removeJdbcConnectionPool(Long id);
 
 
     /**
